corrected uint bug with grid positions

+++ include/imacraft/TerrainGrid.hpp
@@ -9,8 +9,8 @@ namespace imacraft{
 		private:
 			uint8_t *m_data;
 			uint16_t m_width;
-			uint16_t m_northPosition;
-			uint16_t m_eastPosition;
+			int m_northPosition;
+			int m_eastPosition;
 		public:
 			TerrainGrid();
 			~TerrainGrid();
@@ -21,8 +21,8 @@ namespace imacraft{
 			uint32_t length() const;
 			uint16_t width() const;
 			uint16_t height() const;
-			uint16_t getNorthPos() const;
-			uint16_t getEastPos() const;
+			int getNorthPos() const;
+			int getEastPos() const;
 	};
 
 }
